+++ README.md
@@ -1,8 +1,2 @@
 # Front-End React + Vite
 
-```bash
-sudo apt install npm
-
-npm install
-
-npm run dev
