--- setupTests.js
@@ -1,2 +1,0 @@
-// setupTests.js
-import '@testing-library/jest-dom/extend-expect';
