+++ src/components/Button/Button.jsx
@@ -1,5 +1,4 @@
 import './Button.css';
-import React from "react";
 
 //Botón con funcionalidad al hacer click
 function Button ({label, onClick}){
