+++ src/containers/App/App.jsx
@@ -1,18 +1,25 @@
-import './App.css';
-import Login from './components/Login/Login.jsx';
-import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';  // para manejar rutas
-import Home from './components/Home/Home.jsx';
+import "./App.css";
+import Login from "./components/Login/Login.jsx";
+import Start from "./components/SetGame/setgame.jsx";
+import {
+  BrowserRouter as Router,
+  Route,
+  Routes,
+  Navigate,
+} from "react-router-dom"; // para manejar rutas
+import Home from "./components/Home/Home.jsx";
 
-function App () {
-    return (    
-        <Router>  
-            <Routes>
-                <Route path="/" element={<Navigate to="/login"/>} />
-                <Route path="/login" element={<Login/>} />
-                <Route path="/home" element={<Home/>} />
-            </Routes>
-        </Router>
-    );
+function App() {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<Navigate to="/login" />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/home" element={<Home />} />
+        <Route path="/home/create-config" element={<Start />} />
+      </Routes>
+    </Router>
+  );
 }
 
-export default App ;
+export default App;
