+++ src/containers/App/components/Home/Home.jsx
@@ -1,20 +1,26 @@
 import Button from "../../../../components/Button/Button";
-import { useNavigate } from 'react-router-dom'; 
+import { useNavigate } from "react-router-dom";
 import React from "react";
 
 function Home() {
-    const navigate = useNavigate(); 
-   
-    const handleLogOut = () => {
-        navigate('/login');
-    };
-    
-    return ( 
-        <div>
-            <h1>Home</h1>
-            <Button label="Logout" onClick={handleLogOut}/>
-        </div>
-    );
+  const navigate = useNavigate();
+
+  const handleLogOut = () => {
+    navigate("/login");
+  };
+
+  const handleCrearPartida = () => {
+    navigate("/create-config");
+  };
+
+  return (
+    <div>
+      <h1>Home</h1>
+      <Button label="Logout" onClick={handleLogOut} />
+      <h1></h1>
+      <Button label="Crear Partida" onClick={handleCrearPartida} />
+    </div>
+  );
 }
 
 export default Home;
