--- src/hooks/Login/loginHelper.js
@@ -1,3 +1,0 @@
-export const checkInput = (input) => {
-    return input.length > 0 && input.length <= 10;
-};
