+++ src/utils/Constants.js
@@ -2,6 +2,7 @@ export const PORT = 8000;
 
 /** URLS */
 export const LOGIN_URL = `http://localhost:${PORT}/login`;
+export const CONFIG_URL = `http://localhost:${PORT}/home/create-config`;
 
 
 /** MENSAJES */
